protect post with post policy

+++ app/Http/Controllers/PostController.php
@@ -22,17 +22,20 @@ class PostController extends Controller
 
     public function create(): Response
     {
+        $this->authorize("create", Post::class);
         return Inertia::render("Admin/Posts/Create");
     }
 
     public function store(CreatePostRequest $request): RedirectResponse
     {
+        $this->authorize("create", Post::class);
         Post::create($request->validated());
         return to_route("posts.index");
     }
 
     public function edit(Post $post): Response
     {
+        $this->authorize("update", $post);
         return Inertia::render("Admin/Posts/Edit", [
             "post" => new PostResource($post)
         ]);
@@ -40,12 +43,14 @@ class PostController extends Controller
 
     public function update(CreatePostRequest $request, Post $post): RedirectResponse
     {
+        $this->authorize("update", $post);
         $post->update($request->validated());
         return to_route("posts.index");
     }
 
     public function destroy(Post $post): RedirectResponse
     {
+        $this->authorize("delete", $post);
         $post->delete();
         return back();
     }
